requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
amqp==5.3.1
asgiref==3.8.1
asttokens==3.0.0
attrs==25.2.0
beautifulsoup4==4.13.3
billiard==4.2.1
boto3==1.37.13
botocore==1.37.13
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
decorator==5.2.1
Django==5.1.7
django-cors-headers==4.7.0
django-extensions==3.2.3
django-filter==25.1
django-storages==1.14.5
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
drf-spectacular==0.28.0
executing==2.2.0
frozenlist==1.5.0
future==1.0.0
greenlet==3.1.1
idna==3.10
inflection==0.5.1
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jikanpy-v4==1.0.2
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.5.0
lxml==5.3.1
matplotlib-inline==0.1.7
multidict==6.2.0
oauthlib==3.2.2
parso==0.8.4
prompt_toolkit==3.0.50
propcache==0.3.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
pyee==13.0.0
Pygments==2.19.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-ffmpeg==2.0.12
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
requests==2.32.3
requests-oauthlib==1.3.1
rpds-py==0.23.1
s3transfer==0.11.4
shikimori-api==1.0.1
simplejson==3.20.1
six==1.17.0
soupsieve==2.6
SQLAlchemy==2.0.39
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.18.3
